added component names and update dashboard

+++ tornado/kill.py
@@ -6,6 +6,7 @@ from influxdb_client import InfluxDBClient
 from influxdb_client.client.exceptions import InfluxDBError
 from influxdb_client.client.write_api import SYNCHRONOUS
 
+
 URL = os.getenv("INFLUX_URL")
 TOKEN = os.getenv("INFLUX_TOKEN")
 ORG = os.getenv("INFLUX_ORG")
@@ -19,19 +20,28 @@ influxdb_client = InfluxDBClient(url=URL, token=TOKEN, org=ORG)
 write_api = influxdb_client.write_api(write_options=SYNCHRONOUS)
 query_api = influxdb_client.query_api()
 bucket_api = influxdb_client.buckets_api()
+delete_api = influxdb_client.delete_api()
+
+
+def kill_rockoon():
+    delete_api.delete(
+        start="1970-01-01T00:00:00Z",
+        stop=datetime.utcnow().isoformat() + "Z",
+        predicate='',
+        bucket="rockoon",
+    )
 
 def killEmAll():
     buckets = bucket_api.find_buckets().buckets
-    for bucket in buckets:
+    buckets_created_by_user = [bucket for bucket in buckets if bucket.type == "user"]
+    for bucket in buckets_created_by_user:
         try:
+            if bucket.name == "rockoon":
+                continue
             bucket_api.delete_bucket(bucket.id)
         except InfluxDBError as e:
             print(f"InfluxDBError: {e}")
 
-    bucket_api.create_bucket(
-        bucket_name="rockoon",
-        org_id=ORG,
-        description="Bucket created at" + str(datetime.utcnow()),
-    )
+    kill_rockoon()
 
 killEmAll()
